backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
pandas
openpyxl
google-genai
